feat: show spinner during the update of Legacy content
Currently, when a Legacy API request takes too long to get a response, the Legacy
page component does not show anything to indicate the request is in progress. And user
can perform other actions like clicking another button. To improve this, a backdrop
with spinner should be display while the request is in progress, and should disappear
after the page content is updated.