build: disable scope hoisting in Parcel build
Since we upgraded Parcel to v2.13.3, a production issue came up where the use of
'html-react-parser' in SearchResult and the whole New Search UI was crashed. The
issue was most likely due to the scope hoisting done by Parcel. Significant amount
of time has been spent to find a proper solution but all the potential fixes we had
tried did not work. As a result, it has been agreed that we will disable scope hoisting
in 25.1 and then replace Parcel with Vite in 25.2.