feat: ensure Wizard validation correctly executed for cloning Item without Attachment
If an Item is configured to have one Attachment at least, the validation fails to prevent saving a
cloned Item without attachments. This is because the cloned Item's XML still contains the original
Item's Attachment metadata. To fix this, the cloned Item's XML must be updated to exclude  the
original Item's Attachment metadata.