chore(eslint): remove excessive ESLint rules
A couple of ESLint rules that are not in the 'recommended' set are
causing issues on merges and doesn't really align with the team's coding
style. They were initially enabled by an external contributor, so we
agree to remove them.

As well as that, I made some (very minor) tweaks in IntegTester to
address issues raised by jsx-a11y eslint plugin. Although accessibility
is not a concern in the IntegTester, it was easier to do that than to
add excludes etc.