chore(deps:swagger-ui): pin react-dom to v18
This fixes the issue with the Swagger UI failing to render. This was
based on discussions at: https://github.com/swagger-api/swagger-ui/issues/10511

Hopefully in the future this could be removed. As surely it should only
be a temporary fix. But for now it allows us to upgrade to Swagger UI v5
from v4.